image-upscaler-ui/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.2
fastapi-cli==0.0.5
fastjsonschema==2.19.1
filelock==3.15.4
Flask==3.0.3
fonttools==4.48.1
fqdn==1.5.1
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.8
ipykernel==6.29.0
ipython==8.20.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterthemes==0.20.0
kiwisolver==1.4.5
lesscpy==0.15.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==2.1.1
openai==1.13.3
opencv-python==4.11.0.86
overrides==7.6.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.3
pexpect==4.9.0
pi==0.1.2
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==5.29.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.17.2
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.17.1
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.38.4
sympy==1.13.1
tenacity==9.1.2
terminado==0.18.0
tinycss2==1.2.1
toml==0.10.2
torch==2.6.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typer==0.12.5
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.3
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==13.0.1
Werkzeug==3.0.2
wikipedia==1.4.0
